moved styles associated with aws charts to component folders
