+++ app/components/AwsEC2Graphs/AwsEC2Graphs.tsx
@@ -1,8 +1,10 @@
 import React, { useContext, useEffect, useState } from 'react';
-import AwsChart from '../charts/AwsChart';
-import { AwsContext } from '../context/AwsContext';
+import AwsChart from '../../charts/AwsChart';
+import { AwsContext } from '../../context/AwsContext';
+import { stringToColor } from '../../utils';
+import './styles.scss'
 import { CircularProgress } from '@mui/material';
-// import zIndex from '@mui/styles/zIndex';
+
 
 const AwsEC2Graphs: React.FC = React.memo(props => {
   const { awsData, setAwsData, isLoading, setLoadingState } = useContext(AwsContext);
@@ -13,24 +15,6 @@ const AwsEC2Graphs: React.FC = React.memo(props => {
       setLoadingState(true);
     };
   }, []);
-
-  const stringToColor = (string: string, recurses = 0) => {
-    if (recurses > 20) return string;
-    function hashString(str: string) {
-      let hash = 0;
-      for (let i = 0; i < str.length; i++) {
-        hash = str.charCodeAt(i) + ((hash << 5) - hash);
-      }
-      let colour = '#';
-      for (let i = 0; i < 3; i++) {
-        const value = (hash >> (i * 8)) & 0xff;
-        colour += `00${value.toString(16)}`.substring(-2);
-      }
-
-      // console.log(colour);
-      return colour;
-    }
-  };
 
   return (
     <div className="charts">
