+++ app/components/AwsEC2Graphs/styles.scss
@@ -1,0 +1,10 @@
+.charts {
+  display: grid;
+  grid-template-columns: auto auto;
+  padding: 10px;
+}
+
+.chart {
+  margin: 10px;
+  text-align: center;
+}
