+++ app/components/AwsECSClusterGraphs.tsx
@@ -1,5 +1,6 @@
 import React, { useContext, useEffect } from 'react';
 import AwsChart from '../charts/AwsChart';
+import { stringToColor } from '../utils';
 import { AwsContext } from '../context/AwsContext';
 
 const AwsECSClusterGraphs: React.FC = React.memo(props => {
@@ -11,24 +12,6 @@ const AwsECSClusterGraphs: React.FC = React.memo(props => {
       setLoadingState(true);
     };
   }, []);
-
-  const stringToColor = (string: string, recurses = 0) => {
-    if (recurses > 20) return string;
-    function hashString(str: string) {
-      let hash = 0;
-      for (let i = 0; i < str.length; i++) {
-        hash = str.charCodeAt(i) + ((hash << 5) - hash);
-      }
-      let colour = '#';
-      for (let i = 0; i < 3; i++) {
-        const value = (hash >> (i * 8)) & 0xff;
-        colour += `00${value.toString(16)}`.substring(-2);
-      }
-
-      // console.log(colour);
-      return colour;
-    }
-  };
 
   const activeServices = Object.keys(awsEcsData)
     .slice(1)
